+++ ali/common/src/main/java/com/yanny/ali/network/AbstractServer.java
@@ -56,7 +56,6 @@ import org.jetbrains.annotations.Unmodifiable;
 import org.slf4j.Logger;
 
 import java.util.*;
-import java.util.function.Function;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -509,7 +508,7 @@ public abstract class AbstractServer {
     private static Map<Identifier, IDataNode> processTrades(AliServerRegistry serverRegistry, AliConfig config) {
         Map<Identifier, IDataNode> nodes = new HashMap<>();
 
-        for (Map.Entry<Identifier, Function<IServerUtils, Int2ObjectMap<TradeLevel>>> entry : serverRegistry.getTrades().entrySet()) {
+        for (Map.Entry<Identifier, AliServerRegistry.Trades> entry : serverRegistry.getTrades().entrySet()) {
             Identifier location = entry.getKey();
 
             TooltipContext.set(location);