+++ alicompat/neoforge/src/compat/occultism/java/com/yanny/alicompat/compat/occultism/ItemTradeAccessor.java
@@ -1,42 +1,45 @@
 package com.yanny.alicompat.compat.occultism;
 
 import com.klikli_dev.occultism.common.entity.spirit.wonderingtrader.WonderingTrades;
+import com.mojang.datafixers.util.Either;
+import com.yanny.aci.api.RangeValue;
+import com.yanny.aci.tooltip.TooltipNode;
+import com.yanny.ali.api.IDataNode;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.plugin.common.trades.ItemsToItemsNode;
 import com.yanny.alicompat.accessor.BaseAccessor;
 import com.yanny.alicompat.accessor.FieldAccessor;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.ItemStackTemplate;
-import net.minecraft.world.item.trading.TradeCost;
-import net.minecraft.world.item.trading.VillagerTrade;
-import org.jetbrains.annotations.NotNull;
 
 public class ItemTradeAccessor extends BaseAccessor<WonderingTrades.ItemTrade> {
     @FieldAccessor
     private ItemStack input;
-
     @FieldAccessor
     private int maxUses;
-
     @FieldAccessor
     private int villagerXp;
-
     @FieldAccessor
     private ItemStack result;
-
-    @FieldAccessor
-    private float priceMultiplier;
 
     public ItemTradeAccessor(WonderingTrades.ItemTrade parent) {
         super(parent);
     }
 
-    @NotNull
-    public VillagerTrade getTrade() {
-        return VillagerTrade.builder(
-                new TradeCost(input.getItem(), input.getCount()),
-                ItemStackTemplate.fromNonEmptyStack(result),
-                maxUses,
-                villagerXp,
-                priceMultiplier
-        ).build();
+    public IDataNode getNode(IServerUtils utils, TooltipNode condition) {
+        return new ItemsToItemsNode(
+                utils,
+                Either.left(input.copyWithCount(1)),
+                new RangeValue(input.getCount()),
+                TooltipNode.empty(),
+                Either.left(ItemStack.EMPTY),
+                new RangeValue(1),
+                TooltipNode.empty(),
+                Either.left(result.copyWithCount(1)),
+                new RangeValue(result.getCount()),
+                TooltipNode.empty(),
+                new RangeValue(maxUses),
+                new RangeValue(villagerXp),
+                condition
+        );
     }
 }